Graphics/unified mode shader fixes 3
Part of [POI-899](https://jira.unity3d.com/browse/POI-899).

Fix float to min16float SRP shader warnings that appear after applying d1cdd8bc79f707d6d25e9bf1953504fca2e4c5b5 and setting Shader Precision Model to Unified. There are too many warnings to fix in one PR, so doing smaller PRs that can be easily tested and reviewed to avoid breaking anything.

+++ Packages/com.unity.render-pipelines.core/Runtime/Lighting/ProbeVolume/ProbeVolumeBlendStates.hlsl
@@ -4,7 +4,7 @@
 #include "Packages/com.unity.render-pipelines.core/ShaderLibrary/Common.hlsl"
 #include "Packages/com.unity.render-pipelines.core/Runtime/Lighting/ProbeVolume/ProbeVolume.hlsl"
 
-APVSample BlendAPVSamples(APVSample state0, APVSample state1, float factor)
+APVSample BlendAPVSamples(APVSample state0, APVSample state1, half factor)
 {
     APVSample result;
     result.L0   = lerp(state0.L0,   state1.L0,   factor);
